Move MetaTest to execution package
